refactor(GameOverState): reimplement GameOverState without UI framework
- Remove Stage, Skin, and UI components dependency
- Simplify render method for game over screen
- Use basic BitmapFont for text rendering
- Handle font resource disposal manually

+++ core/src/main/java/game/mygame/entities/ZigzagEnemy.java
@@ -33,6 +33,6 @@ public class ZigzagEnemy extends Enemy {
 
     @Override
     public EnemyPrototype clone() {
-        return new ZigzagEnemy(x, y, speed, health, scoreValue, texture);
+        return new ZigzagEnemy(x, y, speed, (int) health, scoreValue, texture);
     }
 }
